Doxygenated Label

+++ include/tobkit/label.h
@@ -26,25 +26,57 @@
 namespace TobKit
 {
 
+/**
+ * A Label that shows some text on the screen.
+ * \image html label.png
+ */
 class Label: public Widget {
 
 	public:
+        /**
+         * Creates a Label
+         * \param owner the GUI that the Label belongs to
+         * \param caption the text on the Label
+         * \param x x-position on the screen
+         * \param y y-position on the screen
+         * \param has_border whether the Label is drawn with a border (This sugests you can edit it, so if you give the Label a border, be sure to bring up a keyboard when it throws signal_pushed.)
+         * \param width width of the Label
+         * \param height height of the Label
+         * \param visible if the Label is drawn and responds to input
+         */
 	    Label(WidgetManager *owner, const string &caption, int x, int y,
 	            bool has_border=false, int width=-1, int height=-1,
 	            bool visible=true);
 
 		~Label(void);
 
-		// Drawing request
+		/**
+		 * Request drawing the Label explicitly.
+		 */
 		void pleaseDraw(void);
 
-		// Event calls
-		void penDown(u8 x, u8 y);
+		/**
+         * The pen is put down on the Widget.
+         * This method is called by the GUI.
+         * You only need this method if you want to simulate the event manually.
+         */
+		void penDown(int x, int y);
 
+		/*
+         * Change the Widget's caption
+         * \param caption the new caption.
+         */
 		void setCaption(const string &caption);
+
+		/**
+		 * Get the Widget's caption.
+		 * \returns the caption of the Widget
+		 */
 		const string &getCaption(void);
 
-		// Signals
+		/**
+		 * This signal is thrown once the pen touches the Widget.
+		 */
 		sigc::signal<void> signal_pushed;
 
 	private:
